Add Float and Mat4 uniforms, GLCallAssign macro
This commit adds the ability to set float and Mat4 uniforms for the
shaders.

An example of this change is made in the HellowYellow application and
the OnlyPosition shader.

This commit also adds the GLCallAssign macro, which is a version of the
GLCall macro that can be used for assignment. The macro is used as such:

GLCallAssign(int a, someFunctionReturningInt());