Add Scene, Materials, and Transforms
This commit adds a first pass at scene object management and iteration.
For now the Scene class has a list of objects in the scene.

This commit also brings a VERY barebones version of materials,
consisting exclusively of which shader program to use.

A transform class is introduced to combine the position and scale with a
function to convert these into a model matrix. Rotation is still
missing.

Issues: #5