Convert to CMAKE, Basic asset file management
This commit transforms the project from being MS Build based to being
CMake based to support not using Visual Studio.

A VERY basic file management system is also implemented to allow
shaders/other resources to be loaded with paths relative to a common
asset folder. This asset folder can be placed anywhere from the exe and
up. Eg if the .exe is in folder 'game/build/release/game.exe' the asset
folder can be in any of the current or above folders.

Resolves #8