Add AllocConsole, revert shader, move triangle
AllocConsole has been added and the various in/outputs are redirected to
it. It currently calls FreeConsole() before allocating a console, which
works well in Visual Studio Code, but maybe cause problems on other
systems.

The shader now interpolates triangle colors again, instead of the flat
red shading. In addition, one of the triangles now move back and forth
as the frame runs to ensure that matrices and uniforms are correctly
updated every frame.

Resolves #9