Added initial Intel VTune profiler project
In search of a profiler I found Intel VTune, which might satisfy my CPU profiling needs. It would be nice to see if I can find a program that combines memory profiling and CPU profiling (Sorta like VS :unamused:). But for now I will give this a try and use a separate tool for memory debugging. It's important to note that external memory debugging probably wont be that important, as I don't intend to use runtime memory allocation past initialization. But still good be sure.