Fix lighting, uniform safety, deltatime, quad fix
Lighting was using wrong euler rotations. We now just use glm::rotateX
and rotateY based on yaw and pitch. In addition, lighting calulations
were using normals in object-space, while the light direction was in
world-space. This is fixed by adding model matrix to the shader
pipeline, and adding an out/in variable pair in the shaders for world
normals.

Uniforms used to crash whenever a location was not found. This turns out
to be annoying when variables get optimized out, so it was in turn taken
out.

We now use Glfw's glfwGetTime() to implement a m_deltaTime fiew in
Application to make animations more good.

The quad had a normal pointing upwards (0, 1 , 0) which is wrong as
heck, since the quad by default is facing us. This is fixed by giving
quad a normal of (0, 0, 1).

Resolves: #28, #26, #25